requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
Werkzeug==3.1.3
